[Fix] 파일구조 및 깔끔하게 변경함
