--- src/index.js
@@ -1,7 +1,0 @@
-import "dotenv/config";
-import { createApp } from "./app.js";
-const PORT = process.env.PORT || 3000;
-const app = createApp();
-app.listen(PORT, () =>
-  console.log(`Server listening on http://localhost:${PORT}`),
-);
